[#6 - #19 - #20] - validation: improve SOC inside the validator and column computation algorithm
