+++ lib/util/columnComputation.js
@@ -15,7 +15,7 @@ function bigger(a, b) {
 }
 
 module.exports = {
-  computeAppropriateColumn: function (items, context) {
+  computeAppropriateColumn: function (items, context, wrongColumnHandler) {
     var source = context.getSourceCode();
     var result = items.reduce(function (prev, cur) {
       var curCol = cur.loc.start.column;
@@ -26,20 +26,7 @@ module.exports = {
     }, null);
 
     if (null === result) {
-      var actual = this.computeActualColumn(items, source);
-      items.forEach(function (token) {
-        context.report({
-          message: 'The punctuator column must be placed directly after the longest variable!',
-          node:    token,
-          fix:     function (fixer) {
-            var before = source.getTokenBefore(token);
-            return fixer.removeRange([
-              before.range[1] + (actual - before.loc.end.column),
-              token.range[0]
-            ]);
-          }
-        });
-      });
+      wrongColumnHandler();
       return;
     }
