+++ lib/util/validatePunctuators.js
@@ -11,35 +11,14 @@
 'use strict';
 
 var columnComputation = require('./columnComputation');
-
-function repeat(val, times) {
-  var buffer = '';
-  for (var i = 0; i < times; i++) {
-    buffer += val;
-  }
-
-  return buffer;
-}
-
-function getWrongAlignmentFixer(column, cur, punctuator, source) {
-  return function(fixer) {
-    if (column > cur) {
-      return fixer.insertTextBefore(punctuator, repeat(' ', column - cur));
-    }
-    var before = source.getTokenBefore(punctuator);
-    return fixer.removeRange([
-      before.range[1] + (column - before.loc.end.column),
-      punctuator.range[0]
-    ]);
-  };
-}
+var fixer = require('./fixer');
 
 module.exports = function(items, context, node, source) {
   if (items.length <= 1) {
     return;
   }
 
-  var column = columnComputation.computeAppropriateColumn(items, context),
+  var column = columnComputation.computeAppropriateColumn(items, context, fixer.getWrongColumnHandler(items, source, context)),
       actual = columnComputation.computeActualColumn(items, source);
   if (typeof column !== 'undefined') {
     items.forEach(function (punctuator) {
@@ -48,7 +27,7 @@ module.exports = function(items, context, node, source) {
         context.report({
           message: 'Invalid indent!',
           node:    punctuator,
-          fix:     getWrongAlignmentFixer(actual, cur, punctuator, source)
+          fix:     fixer.getWrongAlignmentFixer(actual, cur, punctuator, source)
         });
       }
     });
